refactor(generation): move image generation logic to a dedicated module
- Extracted the `generateImage` function from the generation module into a new `imgGeneration.ts` file, promoting modularity and separation of concerns.
- Improved the structure of the image generation logic by encapsulating related functionality, including target size determination and metering event creation, into dedicated functions.
