+++ packages/core/src/index.ts
@@ -26,3 +26,4 @@ export * from "./cache.ts";
 export * from "./ragknowledge.ts";
 export * from "./utils.ts";
 export * from "./tokenTrimming.ts";
+export * from "./imgGeneration.ts";
